+ Lazy Loading for UI Framework to prepare for matching VContainer.

+++ Assets/Scripts/Core/Stats/IStatProvider.cs
@@ -3,6 +3,6 @@ using UnityEngine;
 public interface IStatProvider 
 {
     CharacterConfig BaseConfig { get; }
-    float GetTotalStat(StatType type);
+    int GetTotalStat(StatType type);
     float GetBaseStat(StatType type);
 }
